refactor: Move LegalHoldModal to zustand
