+++ src/script/conversation/MessageRepository.ts
@@ -1224,7 +1224,10 @@ export class MessageRepository {
     if (blockSystemMessage) {
       conversation.blockLegalHoldMessage = true;
     }
-    const missing = await this.conversationService.getAllParticipantsClients(conversation.qualifiedId);
+    const missing = await this.conversationService.getAllParticipantsClients(
+      conversation.qualifiedId.id,
+      conversation.qualifiedId.domain,
+    );
     const deleted = findDeletedClients(missing, await this.generateRecipients(conversation));
     await this.onClientMismatch?.({deleted, missing} as ClientMismatch, conversation, true);
     if (blockSystemMessage) {
