+++ src/script/entity/Conversation.ts
@@ -33,6 +33,7 @@ import ko from 'knockout';
 import {container} from 'tsyringe';
 import {Cancelable, debounce} from 'underscore';
 
+import {useLegalHoldModalState} from 'Components/Modals/LegalHoldModal/LegalHoldModal.state';
 import {t} from 'Util/LocalizerUtil';
 import {getLogger, Logger} from 'Util/Logger';
 import {matchQualifiedIds} from 'Util/QualifiedId';
@@ -57,7 +58,6 @@ import {ConversationVerificationState} from '../conversation/ConversationVerific
 import {NOTIFICATION_STATE} from '../conversation/NotificationSetting';
 import {ConversationError} from '../error/ConversationError';
 import {StatusType} from '../message/StatusType';
-import {useAppMainState} from '../page/state';
 import {ConversationRecord} from '../storage/record/ConversationRecord';
 import {TeamState} from '../team/TeamState';
 
@@ -331,8 +331,8 @@ export class Conversation {
         this.legalHoldStatus(hasLegalHold ? LegalHoldStatus.ENABLED : LegalHoldStatus.DISABLED);
       }
       if (!hasLegalHold) {
-        const {legalHoldModal} = useAppMainState.getState();
-        legalHoldModal.closeRequestModal(this.id);
+        const {closeRequestModal} = useLegalHoldModalState.getState();
+        closeRequestModal(this.id);
       }
       return hasLegalHold;
     });
